# [Profiling] Reduce number of executables queried