This PR reduces the querying of executables to file IDs from native and kernel frames, as only these have entries in `profiling-executables`.

Here is an example (29496 docs queried before and 183 docs queried after).

**Before**
```
[2023-12-11T07:50:18,415][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraces found [16825] stack traces, [55939] frames, [29496] executables.
[2023-12-11T07:50:18,415][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraces took [725.197689 ms].
[2023-12-11T07:50:18,551][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraceDetails found [33079] stack frames, [182] executables.
[2023-12-11T07:50:18,551][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraceDetails took [133.946369 ms].
```

**After**
```
[2023-12-11T08:45:55,818][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraces found [16825] stack traces, [55939] frames, [183] executables.
[2023-12-11T08:45:55,818][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraces took [484.456895 ms].
[2023-12-11T08:45:55,973][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraceDetails found [33079] stack frames, [182] executables.
[2023-12-11T08:45:55,973][DEBUG][o.e.x.p.TransportGetStackTracesAction] [runTask-0] retrieveStackTraceDetails took [153.70309 ms].
```

The drop in latency for `retrieveStackTraces` might be an outlier or due to my local testing with just one node.